+++ src/compenents/people/PeopleItem.js
@@ -1,67 +1,66 @@
 import axios from 'axios';
 import React, { useState } from 'react'
+import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom'
 import { mainUrl, storageUrl } from '../../API';
 import profile from "./../../assets/img/profile.jpg"
 
 function PeopleItem({ user }) {
-    // console.log(user);
+    // =============AUTH================
+    let refreshLogin = useSelector(state => state.refreshLogin);
+    let token = refreshLogin.token;
+    let loadingToken = refreshLogin.loading;
+    // =============================
+    // console.log(user.friend_status);
     let profile_img = user.profile_img === null ? profile : user.profile_img;
     //friends/request/send
-    const [loading, setLoading] = useState(false);
+    const [sendReqLoading, setSendReqLoading] = useState(false);
     const [error, setError] = useState(false);
-    const [reqSend, setReqSend] = useState(false);
+    const [frientStatus, setFriendStatus] = useState(user.friend_status);
+    const [cancelLoading, setCancelLoading] = useState(false);
+
 
     const sendFriendRequest = async () => {
-        setLoading(true);
-        //get auth info
-        let authInfo = JSON.parse(localStorage.getItem('authInfo'));
-        let token = authInfo.token;
+        setSendReqLoading(true);
         let config = {
             headers: {
                 'Authorization': 'Bearer ' + token,
                 'Content-Type': 'multipart/form-data'
             }
         }
-        await axios.post(mainUrl + '/friends/request/send', { "request_to": user.id }, config)
-            .then(info => {
-                if (info.data.success === true) {
-                    setReqSend(true);
-                }
-                setLoading(false);
-                setError(false);
-            })
-            .catch(err => {
-                setLoading(false);
-                setError(true);
-            })
+        if (!loadingToken)
+            await axios.post(mainUrl + '/friends/request/send', { "request_to": user.id }, config)
+                .then(info => {
+                    setFriendStatus(0);
+                    setError(false);
+                })
+                .catch(err => {
+                    setError(true);
+                }).finally(() => {
+                    setSendReqLoading(false);
+                });
     }
 
-    const [cancelLoading, setCancelLoading] = useState(false);
-    const [reqCanceled, setReqCanceled] = useState(false);
+
     const cancelFriendReq = async () => {
         //get auth info
         setCancelLoading(true);
-        let authInfo = JSON.parse(localStorage.getItem('authInfo'));
-        let token = authInfo.token;
         let config = {
             headers: {
                 'Authorization': 'Bearer ' + token,
                 'Content-Type': 'multipart/form-data'
             }
         }
-        await axios.delete(mainUrl + `/friend/${user.id}/delete`, config)
-            .then(info => {
-                console.log(info.data);
-                if (info.data.success === true) {
-                    setReqCanceled(true);
-                }
-            })
-            .catch(err => {
-                console.log(err);
-            }).finally(() => {
-                setCancelLoading(false);
-            })
+        if (!loadingToken)
+            await axios.delete(mainUrl + `/friend/${user.id}/delete`, config)
+                .then(info => {
+                    setFriendStatus(-1);
+                })
+                .catch(err => {
+                    // console.log(err);
+                }).finally(() => {
+                    setCancelLoading(false);
+                })
     }
 
     return (
@@ -79,7 +78,7 @@ function PeopleItem({ user }) {
                         user.breed_name
                     }
                     {
-                        user.friend_status === 1 && <span>&nbsp;&nbsp;(Friend)</span>
+                        frientStatus === 1 && <span>&nbsp;&nbsp;(Friend)</span>
                     }
                 </span>
             </div>
@@ -96,82 +95,57 @@ function PeopleItem({ user }) {
                 </div>
             </div>
             <div className="button-box-freind flex-between w-full">
-                {/* <button className="button-cancel bot-req bot-fr bot-fr1">
-                    Delete&nbsp;
-                    <i className="fa-solid fa-user-slash"></i>
-                </button> */}
+
                 {
-                    user.friend_status === -1 ?
-                        <>
-                            {
-                                reqSend ?
-                                    <button onClick={cancelFriendReq} className="button-ok bot-req bot-fr bot-gro bg-warning">
-                                        cancel Friend request&nbsp;
-                                        <i className="fa-solid fa-plus"></i>
-                                    </button>
-                                    :
-                                    <>
+                    (frientStatus === -1) &&
+                    <>
+                        {
+                            sendReqLoading ?
+                                <button className="button-ok bot-req bot-fr bot-gro" style={{ opacity: '0.5' }}>
+                                    Add friend&nbsp;
+                                    <i className="fas fa-spinner fa-pulse"></i>
+                                </button>
+                                :
+                                <button onClick={sendFriendRequest} className="button-ok bot-req bot-fr bot-gro">
+                                    Add friend&nbsp;
+                                    <i className="fa-solid fa-plus"></i>
+                                </button>
 
-                                        {
-                                            loading ?
-                                                <button className="button-ok bot-req bot-fr" style={{ opacity: '0.5' }}>
-                                                    Add friend&nbsp;
-                                                    <i className="fa-solid fa-user-plus"></i>
-                                                </button>
-                                                :
-                                                <button className="button-ok bot-req bot-fr" onClick={() => sendFriendRequest()}>
-                                                    Add friend&nbsp;
-                                                    <i className="fa-solid fa-user-plus"></i>
-                                                </button>
-                                        }
 
-                                    </>
-
-                            }
-
-                        </>
-                        :
-                        <>
-                            {
-                                user.friend_status === 0 ?
-                                    <>
-                                        {
-                                            !cancelLoading ?
-                                                <>
-                                                    {
-                                                        reqCanceled ?
-                                                            <button className="button-ok bot-req bot-fr" onClick={sendFriendRequest}>
-                                                                Add friend&nbsp;
-                                                                <i className="fa-solid fa-user-plus"></i>
-                                                            </button>
-                                                            :
-                                                            <button className="button-ok bot-req bot-fr bot-gro bg-warning" style={{ opacity: "0.5" }}>
-                                                                cancel Friend request&nbsp;
-                                                                <i className="fa-solid fa-plus"></i>
-                                                            </button>
-                                                    }
-                                                </>
-
-                                                :
-
-                                                <button onClick={cancelFriendReq} className="button-ok bot-req bot-fr bot-gro bg-warning">
-                                                    cancel Friend request&nbsp;
-                                                    <i className="fa-solid fa-plus"></i>
-                                                </button>
-                                        }
-
-                                    </>
-                                    :
-                                    <Link to={`/messanger/${user.id}`}>
-                                        <button className="button-ok bot-req bot-fr">
-                                            Send Message&nbsp;&nbsp;&nbsp;
-                                            <i className="fa-solid fa-paper-plane"></i>
-                                        </button>
-                                    </Link>
-                            }
-                        </>
+                        }
+                    </>
 
                 }
+
+                {
+                    (frientStatus === 0) &&
+                    <>
+                        {
+                            cancelLoading ?
+                                <button className="button-ok bot-req bot-fr bot-gro bg-warning" style={{ opacity: '0.5' }}>
+                                    cancel Friend request&nbsp;
+                                    <i className="fas fa-spinner fa-pulse"></i>
+                                </button>
+                                :
+                                <button onClick={cancelFriendReq} className="button-ok bot-req bot-fr bot-gro bg-warning">
+                                    cancel Friend request&nbsp;
+                                    <i className="fa-solid fa-user-slash"></i>
+                                </button>
+                        }
+
+                    </>
+                }
+
+                {
+                    (frientStatus === 1) &&
+                    <Link to={`/messanger/${user.id}`}>
+                        <button className="button-ok bot-req bot-fr">
+                            Send Message&nbsp;&nbsp;&nbsp;
+                            <i className="fa-solid fa-paper-plane"></i>
+                        </button>
+                    </Link>
+                }
+
 
             </div>
         </div>
