+++ src/compenents/user/FriendItem.js
@@ -1,10 +1,18 @@
 import axios from 'axios';
 import React, { useState } from 'react'
+import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom'
 import { mainUrl, storageUrl } from '../../API';
 import profile from "./../../assets/img/profile.jpg"
 
 function FriendItem(props) {
+
+    // =============AUTH================
+    let refreshLogin = useSelector(state => state.refreshLogin);
+    let token = refreshLogin.token;
+    let loadingToken = refreshLogin.loading;
+    // =============================
+
     let profile_img = props.friend.profile_img === null ? props.friend.profile_img : profile;
     const [showPopUp, setShowPopUP] = useState(false);
     const [deleteFriendLoading, setDeleteFriendLoading] = useState(false);
@@ -16,13 +24,12 @@ function FriendItem(props) {
 
     const deleteFriendFunc = async () => {
         setDeleteFriendLoading(true);
-        let authInfo = JSON.parse(localStorage.getItem('authInfo'));
-        let token = authInfo.token;
         let config = {
             headers: {
                 'Authorization': 'Bearer ' + token
             }
         }
+        if (!loadingToken)
         await axios.delete(`${mainUrl}/friend/${props.friend.user_id}/delete`, config)
             .then(info => {
                 console.log(info.data);
